feat: enforce unique email and add search with pagination

+++ src/main/java/com/kareem/tone/service/StudentService.java
@@ -2,7 +2,9 @@ package com.kareem.tone.service;
 
 import com.kareem.tone.model.Student;
 import com.kareem.tone.repository.StudentRepository;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
+import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
@@ -17,28 +19,47 @@ public class StudentService {
     public List<Student> getAllStudents() {
         return studentRepository.findAll();
     }
-    public Student addStudent(Student student){
+
+    public Student addStudent(Student student) {
+        if (studentRepository.existsByEmail(student.getEmail())) {
+            throw new RuntimeException("Email already in use: " + student.getEmail());
+        }
         return studentRepository.save(student);
     }
 
-    public Student updateStudent(Long id , Student updatedStudent){
-        Student student = studentRepository.findById(id).orElseThrow();
+    public Student updateStudent(Long id, Student updatedStudent) {
+        Student student = studentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Student with id: " + id + " not found"));
+
+        // check email when update
+        if (!student.getEmail().equals(updatedStudent.getEmail())
+                && studentRepository.existsByEmail(updatedStudent.getEmail())){
+            throw new RuntimeException("Email already in use: " + updatedStudent.getEmail());
+        }
+
         student.setName(updatedStudent.getName());
         student.setEmail(updatedStudent.getEmail());
         student.setAge(updatedStudent.getAge());
         return studentRepository.save(student);
     }
-    public void deleteStudent(long id){
+
+    public void deleteStudent(long id) {
+        if (!studentRepository.existsById(id)) {
+            throw new RuntimeException("Student with id: " + id + " not found");
+        }
         studentRepository.deleteById(id);
     }
 
-    public Student getStudentById(long id){
-        boolean exist = studentRepository.existsById(id);
-
-        if (!exist){
-            throw new RuntimeException("Student with id: " + id + " not found");
-        }
-
-        return studentRepository.findById(id).orElseThrow();
+    public Student getStudentById(long id) {
+        return studentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Student with id: " + id + " not found"));
     }
+
+    public Page<Student> searchStudents(String name , Pageable pageable){
+        if (name == null || name.trim().isEmpty()){
+            return studentRepository.findAll(pageable);
+        }
+        return studentRepository.findByNameContainingIgnoreCase(name,pageable);
+    }
+
 }
